+++ tests/corpus/raw-baseline/x64_O0/xxhash32.c
@@ -46,25 +46,25 @@ uint32_t sym__xxhash32(uint64_t RDI_0, uint64_t RSI_0, uint32_t EDX_0)
                     int32_t stack_m76;
                     tmp_11f80_1 = stack_m16;
                     stack_m76 = (int32_t)*(uint32_t*)tmp_11f80_1;
-                    uint64_t RAX_17 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m60), 13);
+                    uint32_t RAX_17 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m60), 13);
                     stack_m60 = (int32_t)((uint64_t)(int32_t)RAX_17 * 0xffffffff9e3779b1U);
                     tmp_11f80_1 = stack_m16;
                     stack_m16 = tmp_11f80_1 + 4;
                     tmp_11f80_1 = stack_m16;
                     stack_m76 = (int32_t)*(uint32_t*)tmp_11f80_1;
-                    uint64_t RAX_24 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m64), 13);
+                    uint32_t RAX_24 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m64), 13);
                     stack_m64 = (int32_t)((uint64_t)(int32_t)RAX_24 * 0xffffffff9e3779b1U);
                     tmp_11f80_1 = stack_m16;
                     stack_m16 = tmp_11f80_1 + 4;
                     tmp_11f80_1 = stack_m16;
                     stack_m76 = (int32_t)*(uint32_t*)tmp_11f80_1;
-                    uint64_t RAX_31 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m68), 13);
+                    uint32_t RAX_31 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m68), 13);
                     stack_m68 = (int32_t)((uint64_t)(int32_t)RAX_31 * 0xffffffff9e3779b1U);
                     tmp_11f80_1 = stack_m16;
                     stack_m16 = tmp_11f80_1 + 4;
                     tmp_11f80_1 = stack_m16;
                     stack_m76 = (int32_t)*(uint32_t*)tmp_11f80_1;
-                    uint64_t RAX_38 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m72), 13);
+                    uint32_t RAX_38 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)stack_m76 * 0xffffffff85ebca77U) + (uint32_t)stack_m72), 13);
                     stack_m72 = (int32_t)((uint64_t)(int32_t)RAX_38 * 0xffffffff9e3779b1U);
                     tmp_11f80_1 = stack_m16;
                     stack_m16 = tmp_11f80_1 + 4;
@@ -78,13 +78,13 @@ uint32_t sym__xxhash32(uint64_t RDI_0, uint64_t RSI_0, uint32_t EDX_0)
                     uint32_t stack_m92;
                     uint32_t stack_m88;
                     int32_t stack_m84;
-                    uint64_t RAX_43 = (uint64_t)sym__rotl32((uint32_t)stack_m60, 1);
+                    uint32_t RAX_43 = sym__rotl32((uint32_t)stack_m60, 1);
                     stack_m92 = (uint32_t)RAX_43;
-                    uint64_t RAX_44 = (uint64_t)sym__rotl32((uint32_t)stack_m64, 7);
+                    uint32_t RAX_44 = sym__rotl32((uint32_t)stack_m64, 7);
                     stack_m88 = (uint32_t)RAX_44 + stack_m92;
-                    uint64_t RAX_47 = (uint64_t)sym__rotl32((uint32_t)stack_m68, 12);
+                    uint32_t RAX_47 = sym__rotl32((uint32_t)stack_m68, 12);
                     stack_m84 = (int32_t)((uint32_t)RAX_47 + stack_m88);
-                    uint64_t RAX_50 = (uint64_t)sym__rotl32((uint32_t)stack_m72, 18);
+                    uint32_t RAX_50 = sym__rotl32((uint32_t)stack_m72, 18);
                     stack_m44 = (uint32_t)RAX_50 + (uint32_t)stack_m84;
                 }
             }
@@ -107,7 +107,7 @@ uint32_t sym__xxhash32(uint64_t RDI_0, uint64_t RSI_0, uint32_t EDX_0)
                     tmp_11f80_1 = stack_m16;
                     uint32_t stack_m80 = *(uint32_t*)tmp_11f80_1;
                     tmp_11f00_27 = stack_m44;
-                    uint64_t RAX_64 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)(int32_t)stack_m80 * 0xffffffffc2b2ae3dU) + tmp_11f00_27), 17);
+                    uint32_t RAX_64 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)(int32_t)stack_m80 * 0xffffffffc2b2ae3dU) + tmp_11f00_27), 17);
                     stack_m44 = (uint32_t)((uint64_t)(int32_t)RAX_64 * 0x27d4eb2f);
                     tmp_11f80_1 = stack_m16;
                     stack_m16 = tmp_11f80_1 + 4;
@@ -123,7 +123,7 @@ uint32_t sym__xxhash32(uint64_t RDI_0, uint64_t RSI_0, uint32_t EDX_0)
                         tmp_11f00_27 = stack_m44;
                         tmp_11f80_1 = stack_m16;
                         uint8_t tmp_11e00_2 = *(uint8_t*)tmp_11f80_1;
-                        uint64_t RAX_73 = (uint64_t)sym__rotl32((uint32_t)((uint32_t)((uint64_t)(int32_t)tmp_11e00_2 * 0x165667b1) + tmp_11f00_27), 11);
+                        uint32_t RAX_73 = sym__rotl32((uint32_t)((uint32_t)((uint64_t)(int32_t)tmp_11e00_2 * 0x165667b1) + tmp_11f00_27), 11);
                         stack_m44 = (uint32_t)((uint64_t)(int32_t)RAX_73 * 0xffffffff9e3779b1U);
                         tmp_11f80_1 = stack_m16;
                         stack_m16 = tmp_11f80_1 + 1;
